requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
dataclasses-json==0.6.0
faiss-cpu==1.7.4
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.37
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.17.2
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.8.0
InstructorEmbedding==1.0.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
langchain==0.0.300
langsmith==0.0.40
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.6
numpy==1.26.0
openai==0.28.0
packaging==23.1
pandas==2.1.1
Pillow==9.5.0
protobuf==4.24.3
pyarrow==13.0.0
pydantic==2.3.0
pydantic_core==2.6.3
pydeck==0.8.1b0
Pygments==2.16.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rfc3986==1.5.0
rich==13.5.3
rpds-py==0.10.3
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.21
streamlit==1.27.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
transformers==4.33.2
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.5
validators==0.22.0
watchdog==3.0.0
yarl==1.9.2
zipp==3.17.0